+++ docs/mkdocs/docs/whatsNew.md
@@ -4,6 +4,7 @@
 ## Version 0.93  
 
 * Issues Resolved:
+    * [#449](https://github.com/jarvisteach/appJar/issues/449) - [updateListBox](/inputWidgets/#set-listboxes) now has a callFunction flag
     * [#442](https://github.com/jarvisteach/appJar/issues/442) - [SpinBoxes](inputWidgets/#spinbox) now call changeFucntion when changed manually  
     * [#437](https://github.com/jarvisteach/appJar/issues/437) - Page discussing [multiple pages](/multiplePages)  
     * [#436](https://github.com/jarvisteach/appJar/issues/436) - [TextAreas](/inputWidgets/#textarea) can now be programmatically cleared and set while disabled